Assignment2/BE/requirements.txt
jaclang>=0.4.0
byllm>=0.1.0
requests>=2.31.0
python-dotenv>=1.0.0
tree-sitter>=0.20.0
gitpython>=3.1.0
streamlit>=1.28.0
pydantic>=2.0.0
markdown>=3.5.0
graphviz>=0.20.0
pygments>=2.15.0
